requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
opencv-python
tensorflow
datasets
Pillow